+++ packages/mantine/src/components/code-editor/CodeEditor.tsx
@@ -17,10 +17,11 @@ import {
     useComponentDefaultProps,
 } from '@mantine/core';
 import {useUncontrolled} from '@mantine/hooks';
-import Editor, {loader} from '@monaco-editor/react';
+import Editor, {loader, useMonaco} from '@monaco-editor/react';
 import {FunctionComponent, useEffect, useState} from 'react';
 
 import {useParentHeight} from '../../hooks';
+import {XML} from './languages/xml';
 
 const useStyles = createStyles((theme) => ({
     root: {},
@@ -40,7 +41,7 @@ interface CodeEditorProps
      *
      * @default 'plaintext'
      */
-    language?: 'plaintext' | 'json' | 'markdown' | 'python';
+    language?: 'plaintext' | 'json' | 'markdown' | 'python' | 'xml';
     /** Default value for uncontrolled input */
     defaultValue?: string;
     /** Value for controlled input */
@@ -103,6 +104,8 @@ export const CodeEditor: FunctionComponent<CodeEditorProps> = (props) => {
         ...others
     } = useComponentDefaultProps('CodeEditor', defaultProps, props);
     const [loaded, setLoaded] = useState(false);
+    const [registered, setRegistered] = useState(false);
+    const monaco = useMonaco();
     const {classes, theme} = useStyles();
     const [_value, handleChange] = useUncontrolled<string>({
         value,
@@ -113,8 +116,8 @@ export const CodeEditor: FunctionComponent<CodeEditorProps> = (props) => {
     const [parentHeight, ref] = useParentHeight();
 
     const loadLocalMonaco = async () => {
-        const monaco = await import('monaco-editor');
-        loader.config({monaco});
+        const monacoInstance = await import('monaco-editor');
+        loader.config({monaco: monacoInstance});
         setLoaded(true);
     };
 
@@ -125,6 +128,13 @@ export const CodeEditor: FunctionComponent<CodeEditorProps> = (props) => {
             setLoaded(true);
         }
     }, []);
+
+    useEffect(() => {
+        if (monaco && language === 'xml' && !registered) {
+            XML.register(monaco);
+            setRegistered(true);
+        }
+    }, [monaco, language]);
 
     const _label = label ? (
         <Input.Label required={required} {...labelProps}>